split tagtransform header into one file per class
